Fix subtle CoffeeScript inconsistency.
Sometimes coffeescript would turn undefined into void 0, and sometimes
it wouldn't, based on where it was and if you actually needed to do it.
Now that we have a scheme layer we can't be inconsistent because
if you do something like

if false
	undefined
	console.log "woo"

and that turns into

(if false

	(console.log "woo"))

well, that's a big problem.