Be more faithful to the CoffeeScript generated JS.
I realized that to do this thing the right way, you have to generate
Scheme that is the literal translation of the JavaScript that
CoffeeScript was generating before. With that in mind, I have introduced
a few new forms to accomodate for this, i.e. (parenthesized), (docomma),
(ternary), etc, which CoffeeScript has used to great effect. I have also
rewritten some of the old code to better follow CoffeeScript's lead.