Introduce macros.
You have two passes through the AST now. The macro-expansion pass and
the code generation pass.

The idea is that you reach into the AST and find all calls to macros,
which you then evaluate. The evaluation will return a new section of the
the tree. You then rewrite that part of the AST with the new tree.