Add `dump`.
This fixes a significant bug in the macro compilation step, as
`console.log` was unwilling to recurse more than 3 or so times into a
list. That means that eventually you'd have [Object] instead of your
actual list, which would break everything. Our `dump` function, on the
other hand, recurses arbitrarily.