Allow for recursive macros.
That is, don't expand a call to a macro from inside that same macro.

This works because when we use the macro to compile stuff, we compile it
down to a function in js that has the same name, so a recursive macro is
just a recursive function that operates on data.